+++ cmd/joshbot/runtime_test.go
@@ -1137,6 +1137,7 @@ func runtimeTestEnvironment(
 			"?sslmode=disable",
 		databasePasswordFileEnvironment: passwordFile,
 		workerIDEnvironment:             "worker-test",
+		webBotAuthModeEnvironment:       webBotAuthModeUnsigned,
 	}
 }